Update sum-of-two-integers.cpp
add (unsigned int) cast to support left shift operation for negative value.

+++ C++/sum-of-two-integers.cpp
@@ -7,7 +7,7 @@ public:
         while (b) {
             int carry = a & b;
             a ^= b;
-            b = carry << 1;
+            b = (unsigned int)carry << 1;
         }
         return a;
     }
